macOS 10.15 beta 1, Xcode 11 beta 1 (iOS, tvOS, watchOS, iOSSupport 13)

+++ iOS/Frameworks/MediaPlayer.framework/MPTimeMarker.h
@@ -15,12 +15,12 @@
     double _time;
     NSString *_title;
     NSURL *_url;
-    int _markerType;
+    long long _markerType;
     NSDictionary *_metadata;
 }
 
 @property(retain, nonatomic) NSDictionary *metadata; // @synthesize metadata=_metadata;
-@property(readonly, nonatomic) int markerType; // @synthesize markerType=_markerType;
+@property(readonly, nonatomic) long long markerType; // @synthesize markerType=_markerType;
 @property(retain, nonatomic) NSURL *url; // @synthesize url=_url;
 @property(retain, nonatomic) NSString *title; // @synthesize title=_title;
 @property(nonatomic) double time; // @synthesize time=_time;
@@ -32,7 +32,7 @@
 @property(readonly, nonatomic) double maxTime;
 @property(readonly, nonatomic) double comparableTime;
 - (id)description;
-- (id)initWithMarkerType:(int)arg1;
+- (id)initWithMarkerType:(long long)arg1;
 
 @end
