+++ iOS/Frameworks/PhotosUI.framework/PUPhotosGridViewControllerSpec.h
@@ -47,7 +47,7 @@
 - (void)configureCollectionViewGridLayout:(id)arg1 forWidth:(double)arg2 safeAreaInsets:(struct UIEdgeInsets)arg3;
 @property(readonly, nonatomic) long long cellBannerTextAlignment;
 @property(readonly, nonatomic) long long cellFillMode;
-@property(readonly, nonatomic) int thumbnailImageFormat;
+@property(readonly, nonatomic) unsigned short thumbnailImageFormat;
 - (unsigned long long)supportedInterfaceOrientations;
 @property(readonly, nonatomic) struct CGSize contentSizeForViewInPopover;
 - (id)init;
