+++ iOS/PrivateFrameworks/AXMediaUtilities.framework/CIImage-AXMExtras.h
@@ -7,11 +7,11 @@
 #import <CoreImage/CIImage.h>
 
 @interface CIImage (AXMExtras)
-- (void)saveToURL:(id)arg1 withOrientation:(long long)arg2 diagnostics:(id)arg3;
+- (void)saveToURL:(id)arg1 withOrientation:(unsigned int)arg2 diagnostics:(id)arg3;
 - (void)writeImageInAllOrientationsToDirectoryAtURL:(id)arg1 diagnostics:(id)arg2;
-- (id)rotatedImageWithInterfaceOrientation:(long long)arg1 isMirrored:(_Bool)arg2 appliedImageOrientation:(long long *)arg3;
-- (id)rotatedImageWithInterfaceOrientation:(long long)arg1 displayOrientation:(long long)arg2 appliedImageOrientation:(long long *)arg3;
-- (long long)_imageOrientationForInterfaceOrientation:(long long)arg1 isMirrored:(_Bool)arg2;
-- (long long)_imageOrientationForInterfaceOrientation:(long long)arg1 displayOrientation:(long long)arg2;
+- (id)rotatedImageWithInterfaceOrientation:(long long)arg1 isMirrored:(_Bool)arg2 appliedImageOrientation:(unsigned int *)arg3;
+- (id)rotatedImageWithInterfaceOrientation:(long long)arg1 displayOrientation:(long long)arg2 appliedImageOrientation:(unsigned int *)arg3;
+- (unsigned int)_imageOrientationForInterfaceOrientation:(long long)arg1 isMirrored:(_Bool)arg2;
+- (unsigned int)_imageOrientationForInterfaceOrientation:(long long)arg1 displayOrientation:(long long)arg2;
 @end
