+++ iOS/PrivateFrameworks/CoreSuggestionsInternals.framework/SGStorageLocation.h
@@ -14,7 +14,7 @@
 
 @interface SGStorageLocation : NSObject <NSCopying, NSSecureCoding, SGEventLocationForGeocode>
 {
-    unsigned long long _locationType;
+    int _locationType;
     NSString *_label;
     NSString *_address;
     NSString *_airportCode;
@@ -34,7 +34,7 @@
 @property(readonly, nonatomic) NSString *airportCode; // @synthesize airportCode=_airportCode;
 @property(readonly, nonatomic) NSString *address; // @synthesize address=_address;
 @property(readonly, nonatomic) NSString *label; // @synthesize label=_label;
-@property(readonly, nonatomic) unsigned long long locationType; // @synthesize locationType=_locationType;
+@property(readonly, nonatomic) int locationType; // @synthesize locationType=_locationType;
 - (void).cxx_destruct;
 - (id)geocodedLocationWithLatitude:(double)arg1 longitude:(double)arg2 accuracy:(double)arg3 handle:(id)arg4;
 - (id)geocodedLocationWithLabel:(id)arg1 address:(id)arg2 latitude:(double)arg3 longitude:(double)arg4 accuracy:(double)arg5 handle:(id)arg6;
@@ -56,9 +56,9 @@
 - (_Bool)isEqualToStorageLocation:(id)arg1;
 - (id)initWithCoder:(id)arg1;
 - (void)encodeWithCoder:(id)arg1;
-- (id)initWithType:(unsigned long long)arg1 label:(id)arg2 address:(id)arg3 airportCode:(id)arg4 latitude:(double)arg5 longitude:(double)arg6 accuracy:(double)arg7 quality:(double)arg8 handle:(id)arg9;
-- (id)initWithType:(unsigned long long)arg1 label:(id)arg2 airportCode:(id)arg3 latitude:(double)arg4 longitude:(double)arg5 accuracy:(double)arg6 quality:(double)arg7;
-- (id)initWithType:(unsigned long long)arg1 label:(id)arg2 address:(id)arg3 airportCode:(id)arg4 accuracy:(double)arg5 quality:(double)arg6;
+- (id)initWithType:(int)arg1 label:(id)arg2 address:(id)arg3 airportCode:(id)arg4 latitude:(double)arg5 longitude:(double)arg6 accuracy:(double)arg7 quality:(double)arg8 handle:(id)arg9;
+- (id)initWithType:(int)arg1 label:(id)arg2 airportCode:(id)arg3 latitude:(double)arg4 longitude:(double)arg5 accuracy:(double)arg6 quality:(double)arg7;
+- (id)initWithType:(int)arg1 label:(id)arg2 address:(id)arg3 airportCode:(id)arg4 accuracy:(double)arg5 quality:(double)arg6;
 
 // Remaining properties
 @property(readonly, copy) NSString *debugDescription;
