+++ iOS/PrivateFrameworks/ScreenReaderCore.framework/SCRCMathIdentifierExpression.h
@@ -9,17 +9,17 @@
 __attribute__((visibility("hidden")))
 @interface SCRCMathIdentifierExpression : SCRCMathSimpleExpression
 {
-    int _fontStyle;
+    long long _fontStyle;
 }
 
-@property(readonly, nonatomic) int fontStyle; // @synthesize fontStyle=_fontStyle;
+@property(readonly, nonatomic) long long fontStyle; // @synthesize fontStyle=_fontStyle;
 - (_Bool)canFormWordStartingWithExpression:(id)arg1;
 - (_Bool)isNaturalSuperscript;
 - (_Bool)_isPrime;
 - (_Bool)isNumber;
 - (id)mathMLString;
-- (id)speakableDescriptionWithSpeakingStyle:(int)arg1 arePausesAllowed:(_Bool)arg2 isPartOfWord:(_Bool)arg3;
-- (id)speakableDescriptionWithSpeakingStyle:(int)arg1 arePausesAllowed:(_Bool)arg2;
+- (id)speakableDescriptionWithSpeakingStyle:(long long)arg1 arePausesAllowed:(_Bool)arg2 isPartOfWord:(_Bool)arg3;
+- (id)speakableDescriptionWithSpeakingStyle:(long long)arg1 arePausesAllowed:(_Bool)arg2;
 - (id)description;
 - (id)initWithDictionary:(id)arg1;
