+++ iOS/PrivateFrameworks/ScreenReaderCore.framework/SCRCMathLongDivisionExpression.h
@@ -16,7 +16,7 @@ __attribute__((visibility("hidden")))
 
 @property(retain, nonatomic) SCRCMathExpression *divisor; // @synthesize divisor=_divisor;
 - (void).cxx_destruct;
-- (id)speakableDescriptionWithSpeakingStyle:(int)arg1 arePausesAllowed:(_Bool)arg2;
+- (id)speakableDescriptionWithSpeakingStyle:(long long)arg1 arePausesAllowed:(_Bool)arg2;
 - (id)initWithDictionary:(id)arg1;
 
 @end
