+++ iOS/PrivateFrameworks/ScreenReaderCore.framework/SCRCMathSimpleExpression.h
@@ -18,7 +18,7 @@ __attribute__((visibility("hidden")))
 - (void).cxx_destruct;
 - (id)latexMathModeDescription;
 - (id)dollarCodeDescriptionWithNumberOfOuterRadicals:(unsigned long long)arg1 treePosition:(id)arg2;
-- (id)speakableDescriptionWithSpeakingStyle:(int)arg1 arePausesAllowed:(_Bool)arg2;
+- (id)speakableDescriptionWithSpeakingStyle:(long long)arg1 arePausesAllowed:(_Bool)arg2;
 - (_Bool)canBeUsedWithBase;
 - (_Bool)isFunctionName;
 - (id)_functionNames;
