+++ iOS/PrivateFrameworks/UIKitCore.framework/UIKBCacheToken_Key.h
@@ -19,7 +19,7 @@ __attribute__((visibility("hidden")))
     int _displayRowHint;
     struct CGSize _size;
     int _state;
-    int _clipCorners;
+    unsigned long long _clipCorners;
     unsigned long long _groupNeighbors;
     _Bool _usesInsets;
     struct UIEdgeInsets _displayInsets;
